requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
idna==2.10
joblib==1.0.1
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.23.1
scipy==1.5.4
six==1.16.0
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
urllib3==1.26.5
